+++ PackageDelivery.Repository.Implementation/Mappers/Core/OfficeRepositoryMapper.cs
@@ -17,7 +17,7 @@ namespace PackageDelivery.Repository.Implementation.Mappers.Core
                 Latitude = input.latitud,
                 Longitude = input.longitud,
                 Address = input.direccion,
-                IdMunicipality = input.idMunicipio
+                IdMunicipality = (long)input.idMunicipio
             };
         }
